llmleg/pyproject.toml
[tool.poetry]
name = "llmleg"
version = "0.1.0"
description = ""
authors = ["Toby <toby.sachs.quintana@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pdfminer-six = "^20221105"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
